create emline_utils
